+++ Network/constants.py
@@ -1,2 +1,10 @@
 SERVER_PORT = 1112
 
+
+# Connection Protocol
+SYN_REQ = "SYN" # Synchronize Request
+SYN_ACK = "ACK" # Synchronize Acknowledged
+
+
+
+# MCP (Maintanance Connection Protocol)
